cleanup
git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@1678 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/netsiddev_builder/NetSIDConnection.java
@@ -171,7 +171,7 @@ public class NetSIDConnection {
 		send(() -> new Mute(sidNum, voice, mute));
 	}
 
-	public void setVolume(byte sidNum, float volume) {
+	public void setVolume(byte sidNum, byte volume) {
 		send(() -> new SetSidLevel(sidNum, volume));
 	}
